Remove that stupid factory pattern from phenotype handler
